requirements.txt
moviepy==1.0.3
google-api-python-client==2.132.0
google-auth==2.33.0
google-auth-oauthlib==1.2.0
requests==2.32.3
Pillow==10.4.0
numpy==1.26.4
